store/tikv:move tikv.Lock to tikv.kv.Lock
