+++ store/tikv/lock_resolver.go
@@ -14,10 +14,8 @@
 package tikv
 
 import (
-	"bytes"
 	"container/list"
 	"context"
-	"encoding/hex"
 	"fmt"
 	"math"
 	"sync"
@@ -42,6 +40,13 @@ const ResolvedCacheSize = 2048
 
 // bigTxnThreshold : transaction involves keys exceed this threshold can be treated as `big transaction`.
 const bigTxnThreshold = 16
+
+type Lock = kv.Lock
+
+// NewLock creates a new *Lock.
+func NewLock(l *kvrpcpb.LockInfo) *Lock {
+	return kv.NewLock(l)
+}
 
 // LockResolver resolves locks and also caches resolved txn status.
 type LockResolver struct {
@@ -152,44 +157,6 @@ var defaultLockTTL uint64 = 3000
 
 // ttl = ttlFactor * sqrt(writeSizeInMiB)
 var ttlFactor = 6000
-
-// Lock represents a lock from tikv server.
-type Lock struct {
-	Key             []byte
-	Primary         []byte
-	TxnID           uint64
-	TTL             uint64
-	TxnSize         uint64
-	LockType        kvrpcpb.Op
-	UseAsyncCommit  bool
-	LockForUpdateTS uint64
-	MinCommitTS     uint64
-}
-
-func (l *Lock) String() string {
-	buf := bytes.NewBuffer(make([]byte, 0, 128))
-	buf.WriteString("key: ")
-	buf.WriteString(hex.EncodeToString(l.Key))
-	buf.WriteString(", primary: ")
-	buf.WriteString(hex.EncodeToString(l.Primary))
-	return fmt.Sprintf("%s, txnStartTS: %d, lockForUpdateTS:%d, minCommitTs:%d, ttl: %d, type: %s, UseAsyncCommit: %t",
-		buf.String(), l.TxnID, l.LockForUpdateTS, l.MinCommitTS, l.TTL, l.LockType, l.UseAsyncCommit)
-}
-
-// NewLock creates a new *Lock.
-func NewLock(l *kvrpcpb.LockInfo) *Lock {
-	return &Lock{
-		Key:             l.GetKey(),
-		Primary:         l.GetPrimaryLock(),
-		TxnID:           l.GetLockVersion(),
-		TTL:             l.GetLockTtl(),
-		TxnSize:         l.GetTxnSize(),
-		LockType:        l.LockType,
-		UseAsyncCommit:  l.UseAsyncCommit,
-		LockForUpdateTS: l.LockForUpdateTs,
-		MinCommitTS:     l.MinCommitTs,
-	}
-}
 
 func (lr *LockResolver) saveResolved(txnID uint64, status TxnStatus) {
 	lr.mu.Lock()
